Creating utilties file to move common analysis for stencil related hw info
