+++ src/HWUtils.h
@@ -4,10 +4,14 @@
 #include <string>
 #include <vector>
 
+#include "IR.h"
+
 namespace Halide {
   namespace Internal {
 
     std::vector<std::string> get_tokens(const std::string &line, const std::string &delimiter);
+    int id_const_value(const Expr e);
+    std::ostream& operator<<(std::ostream& os, const std::vector<std::string>& vec);
   }
 }
 #endif
